+++ test/Sentry.Unity.Tests/UnityEventExceptionProcessorTests.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using NUnit.Framework;
+using Sentry.Unity.Tests.Stubs;
 
 namespace Sentry.Unity.Tests
 {
@@ -27,13 +28,15 @@ namespace Sentry.Unity.Tests
 
             var sentryExceptionFirstFrame = sentryException.Stacktrace!.Frames[0];
             Assert.AreEqual(string.Empty, sentryExceptionFirstFrame.FileName);
+            Assert.AreEqual(string.Empty, sentryExceptionFirstFrame.AbsolutePath);
         }
 
         [Test]
         public void Process_SdkInfo_Correct()
         {
             // arrange
-            var unityEventProcessor = new UnityEventProcessor();
+            var testApplication = new TestApplication();
+            var unityEventProcessor = new UnityEventProcessor(testApplication);
             var sentryEvent = new SentryEvent();
 
             // act
@@ -47,6 +50,41 @@ namespace Sentry.Unity.Tests
             Assert.IsNotNull(package);
             Assert.AreEqual(UnitySdkInfo.PackageName, package!.Name);
             Assert.AreEqual(UnitySdkInfo.Version, package!.Version);
+        }
+
+        [TestCaseSource(nameof(EditorSimulatorValues))]
+        public void Process_EventDeviceSimulator_SetCorrectly(bool isEditor, bool? isSimulator)
+        {
+            // arrange
+            var testApplication = new TestApplication(isEditor);
+            var unityEventProcessor = new UnityEventProcessor(testApplication);
+            var sentryEvent = new SentryEvent();
+
+            // act
+            unityEventProcessor.Process(sentryEvent);
+
+            // assert
+            Assert.AreEqual(sentryEvent.Contexts.Device.Simulator, isSimulator);
+        }
+
+        private static readonly object[] EditorSimulatorValues =
+        {
+            new object[] { true, true },
+            new object[] { false, null! },
+        };
+
+        [Test]
+        public void Process_ServerName_IsNull()
+        {
+            // arrange
+            var unityEventProcessor = new UnityEventProcessor();
+            var sentryEvent = new SentryEvent();
+
+            // act
+            unityEventProcessor.Process(sentryEvent);
+
+            // assert
+            Assert.IsNull(sentryEvent.ServerName);
         }
     }
 }