+++ CHANGELOG.md
@@ -1,55 +1,92 @@
 # Changelog
 
-## Unreleased
+## 0.1.0
 
-- Sentry UnityLogger aligned to Unity Debug API ([#163](https://github.com/getsentry/sentry-unity/pull/163))
+### Features
+
+- Simplified scene breadcrumbs messages ([#197](https://github.com/getsentry/sentry-unity/pull/197))
+- Embedded link.xml in assembly ([#194](https://github.com/getsentry/sentry-unity/pull/194))
+- Transition scene to test scene loading events (breadcrumbs) ([#185](https://github.com/getsentry/sentry-unity/pull/185))
 
 ### Fixes
 
-- SDK version format correction (#120)
-- Auto compression option is part of drop down (no extra checkbox) (#160)
+- Check/create directory before saving ([#196](https://github.com/getsentry/sentry-unity/pull/196))
+- Exclude SentryOptions.json from release package ([#195](https://github.com/getsentry/sentry-unity/pull/195))
+- default env and version ([#199](https://github.com/getsentry/sentry-unity/pull/199))
+- SentryEvent.ServerName forced to 'null' ([#201](https://github.com/getsentry/sentry-unity/pull/201))
+
+## 0.0.14
+
+### Features
+
+- Simulator is set only when Application.isEditor is true ([#190](https://github.com/getsentry/sentry-unity/pull/190))
+- Sentry UnityLogger aligned to Unity Debug API ([#163](https://github.com/getsentry/sentry-unity/pull/163))
+- Scene manager integration for breadcrumbs ([#170](https://github.com/getsentry/sentry-unity/pull/170))
+
+### Fixes
+
+- Flag simulator based on Application.isEditor ([#184](https://github.com/getsentry/sentry-unity/pull/184))
+- il2cpp remove zeroes from path ([#179](https://github.com/getsentry/sentry-unity/pull/179))
+- SDK version format correction ([#120](https://github.com/getsentry/sentry-unity/pull/120))
+- Auto compression option is part of drop down (no extra checkbox) ([#160](https://github.com/getsentry/sentry-unity/pull/160))
+- Rename DiagnosticsLogger to DiagnosticLogger ([#168](https://github.com/getsentry/sentry-unity/pull/168))
+- SentryOptions config proper check ([#176](https://github.com/getsentry/sentry-unity/pull/176))
+- Diagnostic logger writes to console that it was disabled ([#183](https://github.com/getsentry/sentry-unity/pull/183))
 
 ## 0.0.13
 
-- Missing meta files warnings (#146)
+### Fixes
+
+- Missing meta files warnings ([#146](https://github.com/getsentry/sentry-unity/pull/146))
 
 ## 0.0.12
+
+### Fixes
 
 - Release process improvements
 
 ## 0.0.11
 
+### Feature
+
 - Craft Release
 
 ## 0.0.10
+
+### Fixes
 
 - Add missing meta files (cd9c7fd)
 
 ## 0.0.9
 
-- UPM meta updated (#124)
-- Bump dotnet 3.3.4 (#132)
+### Features
+
+- Unity Sentry SDK programmatic setup ([#130](https://github.com/getsentry/sentry-unity/pull/130))
+  - SentryWindow updated
+
+### Fixes
+
+- UPM meta updated ([#124](https://github.com/getsentry/sentry-unity/pull/124))
+- Bump dotnet 3.3.4 ([#132](https://github.com/getsentry/sentry-unity/pull/132))
   - https://github.com/getsentry/sentry-dotnet/blob/main/CHANGELOG.md#334
   - Bug fixes for performance monitoring
   - Ability to keep failed envelopes for troubleshooting when they are too large
-- Unity Sentry SDK programmatic setup (#130)
-  - SentryWindow updated
 
 ## 0.0.8
 
 ### Fixes
 
-- Bump sentry-dotnet to 8ca0906 - fix IL2CPP crash (#122) 
-- IL2CPP players crash due to startup time detection (#123)
+- Bump sentry-dotnet to 8ca0906 - fix IL2CPP crash ([#122](https://github.com/getsentry/sentry-unity/pull/122))
+- IL2CPP players crash due to startup time detection ([#123](https://github.com/getsentry/sentry-unity/pull/123))
 
 ## 0.0.7
 
 ### Features
 
-- Strip zeroes for ill2cpp builds (#108)
-- Proper sdk name reporting for sentry event (#111)
-- Bump .NET SDK to 3.3.1 (#115)
-- Release package samples (#113)
+- Strip zeroes for ill2cpp builds ([#108](https://github.com/getsentry/sentry-unity/pull/108))
+- Proper sdk name reporting for sentry event ([#111](https://github.com/getsentry/sentry-unity/pull/111))
+- Bump .NET SDK to 3.3.1 ([#115](https://github.com/getsentry/sentry-unity/pull/115))
+- Release package samples ([#113](https://github.com/getsentry/sentry-unity/pull/113))
 
 ## 0.0.6